Improve RetroFrame performance by memcpy into a QImage

See merge request seshpenguin/kretro!6

+++ src/app.h
@@ -37,7 +37,7 @@ public:
     Q_INVOKABLE void stopRetroCore();
     Q_INVOKABLE void resetRetroCore();
 
-    void videoRefresh(const void *data, unsigned width, unsigned height, int pitch);
+    void videoRefresh(const void *data, unsigned width, unsigned height, size_t pitch);
     void audioRefresh(const int16_t *data,size_t frames);
